cleanup
for some reason, i keep getting redirected to to some menutafelalt.aspx or so,
which sucks even harder than the normal version, and i can't seem to get the fcking menu list.
something's off here, it may have to do with not having localstorage or so? cookie limitations,
that sort of thing?